65 yaş check

+++ my-app/src/Test/Pages/TakeTest.js
@@ -1,4 +1,7 @@
 import { useEffect, useState } from "react";
+import {toast} from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
+import { useNavigate } from "react-router-dom";
 import {
   getAllPartsAction,
   getAllQuestionsAction,
@@ -9,15 +12,15 @@ import PartInformation from "../Components/PartInformation";
 import QuestionBody from "../Components/QuestionBody";
 import "../Styles/TakeTest.css";
 
+
 const TakeTest = () => {
   const [id, setId] = useState(null);
   const [currentAssessmentSession, setCurrentAssessmentSession] = useState(0);
   const [incomingAnswer, setIncomingAnswer] = useState("");
   const [qID, setQid] = useState();
 
-  // useEffect(async() =>{
-  //   console.log(incomingAnswer, "answer in parent");
-  // }, [incomingAnswer]);
+  const navigate = useNavigate();
+  toast.configure()
 
   const createAssessmentSession = async () => {
     var jsonData = {
@@ -29,17 +32,18 @@ const TakeTest = () => {
     };
     const a = await createAssessmentSessionAction(jsonData);
     setCurrentAssessmentSession(a);
+    return a;
   };
 
   useEffect(async () => {
     setId(sessionStorage.getItem("userId"));
-    if (id !== null && id !== "") {
-      // null check / "" check
-      // await createAssessmentSession();
-      setTimeout(() => {
-        createAssessmentSession();
-      }, 300);
-    }
+    // if (id !== null && id !== "") {
+    //   // null check / "" check
+    //   // await createAssessmentSession();
+    //   setTimeout(() => {
+    //     createAssessmentSession();
+    //   }, 300);
+    // }
   }, [id]);
 
   const [parts, setParts] = useState([]);
@@ -152,7 +156,7 @@ const TakeTest = () => {
   useEffect(async () => {
     if (
       partIndex === lastPartCount - 1 &&
-      questionIndex === lastQuestionCount - 1
+      questionIndex === lastQuestionCount - 1 && showQuestions
     ) {
       setIsSubmitEnabled(true);
     } else {
@@ -163,27 +167,48 @@ const TakeTest = () => {
   const [completed, setCompleted] = useState(false);
 
   const saveToDb = async () => {
-    const arr = [];
-    for (let index = 0; index < questions.length; index++) {
-      const element = [];
-      const qid = questions[index][0];
-      const ans = localStorage.getItem(qid);
-      element.push(qid);
-      element.push(ans);
-      arr.push(element);
+    let sessionId = currentAssessmentSession;
+    if (id !== null && id !== "") {
+        sessionId = await createAssessmentSession();
     }
-    var jsonData = {
-      data: [
-        {
-          AssessmentSessionId: currentAssessmentSession,
-          AnswerList: arr,
-        },
-      ],
-    };
-    const a = await uploadUserAnswersAction(jsonData);
-    localStorage.clear();
-    setCompleted(true);
+
+    setTimeout(() => {
+      const arr = [];
+      for (let index = 0; index < questions.length; index++) {
+        const element = [];
+        const qid = questions[index][0];
+        const ans = localStorage.getItem(qid);
+        element.push(qid);
+        element.push(ans);
+        arr.push(element);
+      }
+      var jsonData = {
+        data: [
+          {
+            AssessmentSessionId: sessionId,
+            AnswerList: arr,
+          },
+        ],
+      };
+      uploadUserAnswersAction(jsonData);
+      localStorage.clear();
+      setCompleted(true);
+        }, 300);
+
   };
+
+
+  useEffect( async () => {
+    const response = localStorage.getItem(1);
+    if (response) {
+      if (parseInt(response) < 65) {
+        toast.warning('Testi alabilmeniz için yaşınız 65 veya üzeri olmalıdır!',
+           {position: toast.POSITION.TOP_CENTER, autoClose:5000})
+        navigate("/testInformation");
+        localStorage.clear();
+      }
+    }
+  })
 
   return (
     <div className="testPageLayout">
